pyproject.toml
[project]
name = "athena"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "google-generativeai>=0.8.5",
    "piper-tts>=1.3.0",
    "python-dotenv>=1.1.1",
    "sounddevice>=0.5.2",
    "vosk>=0.3.45",
]

[dependency-groups]
dev = [
    "black>=25.9.0",
]
